Safari-609.1.11.1/LayoutTests/editing/style/preserve-selection-direction-expected.txt

Bug 110979: Selection direction is not preserved when applying styles

This test verifies that selection direction is preserved when applying styles.

This is some sample text
selection.collapse(editable.firstChild, 3);
selection.extend(editable.firstChild, 8);
document.execCommand('foreColor', false, 'green');
PASS selection.anchorOffset is 0
PASS selection.focusOffset is 5
editable.innerHTML = 'This is some sample text';
selection.collapse(editable.firstChild, 8);
selection.extend(editable.firstChild, 3);
document.execCommand('foreColor', false, 'green');
PASS selection.anchorOffset is 5
PASS selection.focusOffset is 0
PASS successfullyParsed is true

TEST COMPLETE
